(refactor) Move the call to drop and create the db to the index file.
Moves the server utility functions to it's own folder and file on the server to better show separation of ideas between the utils functions and the socket paths.
