+++ server/utilities/utilityFunctions.js
